+++ src/asm.rs
@@ -2,12 +2,14 @@
 use crate::asm::statements::Label;
 use crate::cached_lines::CachedLines;
 use crate::demangle::LabelKind;
+pub use crate::sources::Source;
+use crate::sources::{locate_sources, path_formatter};
 use crate::{
-    CallGraph, Dumpable, Item, RawLines, URange, color, demangle, esafeprintln, get_context_for,
+    CallGraph, Dumpable, Item, RawLines, URange, color, demangle, get_context_for,
     safeprintln,
 };
 // TODO, use https://sourceware.org/binutils/docs/as/index.html
-use crate::opts::{Format, NameDisplay, RedundantLabels, SourcesFrom};
+use crate::opts::{Format, NameDisplay, RedundantLabels};
 
 mod statements;
 
@@ -15,11 +17,10 @@ use nom::Parser as _;
 use owo_colors::OwoColorize;
 pub use statements::{Directive, GenericDirective, Instruction, Statement};
 use statements::{Loc, parse_statement};
-use std::borrow::Cow;
 use std::cell::RefCell;
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 use std::ops::Range;
-use std::path::{Display, Path, PathBuf};
+use std::path::{Path, PathBuf};
 
 type SourceFile = (String, Option<(Source, CachedLines)>);
 
@@ -409,186 +410,6 @@ fn dump_range(
     }
 
     Ok(())
-}
-
-/// Returns a closure that trims the paths
-fn path_formatter() -> impl for<'p> Fn(&'p Path, &'p mut PathBuf) -> Display<'p> {
-    let current_dir = std::env::current_dir().unwrap_or_default();
-    let home_dir = std::env::home_dir();
-    let home = if std::path::MAIN_SEPARATOR == '/' {
-        "~"
-    } else {
-        "%userprofile%"
-    };
-    move |path, tmp| {
-        if path.is_absolute() {
-            if let Ok(rel) = path.strip_prefix(&current_dir) {
-                rel
-            } else if let Some(path_in_home) = home_dir
-                .as_ref()
-                .and_then(|home| path.strip_prefix(home).ok())
-            {
-                tmp.clear();
-                tmp.push(home);
-                tmp.push(path_in_home);
-                &*tmp
-            } else {
-                path
-            }
-        } else {
-            path
-        }
-        .display()
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum Source {
-    Crate,
-    External,
-    Stdlib,
-    Rustc,
-}
-
-impl Source {
-    fn show_for(&self, from: SourcesFrom) -> bool {
-        match self {
-            Source::Crate => true,
-            Source::External => match from {
-                SourcesFrom::ThisWorkspace => false,
-                SourcesFrom::AllCrates | SourcesFrom::AllSources => true,
-            },
-            Source::Rustc | Source::Stdlib => match from {
-                SourcesFrom::ThisWorkspace | SourcesFrom::AllCrates => false,
-                SourcesFrom::AllSources => true,
-            },
-        }
-    }
-}
-
-// DWARF information contains references to source files
-// It can point to 3 different items:
-// 1. a real file, cargo-show-asm can just read it
-// 2. a file from rustlib, sources are under $sysroot/lib/rustlib/src/rust/$suffix
-//    Some examples:
-//        /rustc/a55dd71d5fb0ec5a6a3a9e8c27b2127ba491ce52/library/core/src/iter/range.rs
-//        /private/tmp/rust-20230325-7327-rbrpyq/rustc-1.68.1-src/library/core/src/option.rs
-//        /rustc/cc66ad468955717ab92600c770da8c1601a4ff33\\library\\core\\src\\convert\\mod.rs
-// 3. a file from prebuilt (?) hashbrown, sources are probably available under
-//    cargo registry, most likely under ~/.cargo/registry/$suffix
-//    Some examples:
-//        /cargo/registry/src/github.com-1ecc6299db9ec823/hashbrown-0.12.3/src/raw/bitmask.rs
-//        /Users/runner/.cargo/registry/src/github.com-1ecc6299db9ec823/hashbrown-0.12.3/src/map.rs
-// 4. rustc sources:
-//    /rustc/89e2160c4ca5808657ed55392620ed1dbbce78d1/compiler/rustc_span/src/span_encoding.rs
-//    $sysroot/lib/rustlib/rust-src/rust/compiler/rustc_span/src/span_encoding.rs
-fn locate_sources(sysroot: &Path, workspace: &Path, path: &Path) -> Option<(Source, PathBuf)> {
-    let mut path = Cow::Borrowed(path);
-    // a real file that simply exists
-    if path.exists() {
-        let source = if path.starts_with(workspace) {
-            Source::Crate
-        } else {
-            Source::External
-        };
-
-        return Some((source, path.into()));
-    }
-
-    let no_rust_src = || {
-        esafeprintln!(
-            "You need to install rustc sources to be able to see the rust annotations, try\n\
-                                       \trustup component add rust-src"
-        );
-        std::process::exit(1);
-    };
-
-    // then during crosscompilation we can get this cursed mix of path names
-    //
-    // /rustc/cc66ad468955717ab92600c770da8c1601a4ff33\\library\\core\\src\\convert\\mod.rs
-    //
-    // where one bit comes from the host platform and second bit comes from the target platform
-    // This feels like a problem in upstream, but supporting that is not _that_ hard.
-    //
-    // I think this should take care of Linux and MacOS support
-    if (path.starts_with("/rustc/") || path.starts_with("/private/tmp"))
-        && path
-            .as_os_str()
-            .to_str()
-            .is_some_and(|s| s.contains('\\') && s.contains('/'))
-    {
-        let cursed_path = path
-            .as_os_str()
-            .to_str()
-            .expect("They are coming from a text file");
-        path = Cow::Owned(PathBuf::from(cursed_path.replace('\\', "/")));
-    }
-
-    // /rustc/89e2160c4ca5808657ed55392620ed1dbbce78d1/compiler/rustc_span/src/span_encoding.rs
-    if path.starts_with("/rustc") && path.iter().any(|c| c == "compiler") {
-        let mut source = sysroot.join("lib/rustlib/rustc-src/rust");
-        for part in path.components().skip(3) {
-            source.push(part);
-        }
-
-        if source.exists() {
-            return Some((Source::Rustc, source));
-        } else {
-            no_rust_src();
-        }
-    }
-
-    // rust sources, Linux style
-    if path.starts_with("/rustc/") {
-        let mut source = sysroot.join("lib/rustlib/src/rust");
-        for part in path.components().skip(3) {
-            source.push(part);
-        }
-        if source.exists() {
-            return Some((Source::Stdlib, source));
-        } else {
-            no_rust_src();
-        }
-    }
-
-    // rust sources, MacOS style
-    if path.starts_with("/private/tmp") && path.components().any(|c| c.as_os_str() == "library") {
-        let mut source = sysroot.join("lib/rustlib/src/rust");
-        for part in path.components().skip(5) {
-            source.push(part);
-        }
-        if source.exists() {
-            return Some((Source::Stdlib, source));
-        } else {
-            no_rust_src();
-        }
-    }
-
-    // cargo registry, Linux and macOS look for cargo/registry and .cargo/registry
-    if let Some(ix) = path
-        .components()
-        .position(|c| c.as_os_str() == "cargo" || c.as_os_str() == ".cargo")
-        .and_then(|ix| path.components().nth(ix).zip(Some(ix)))
-        .and_then(|(c, ix)| (c.as_os_str() == "registry").then_some(ix))
-    {
-        // It does what I want as far as *nix is concerned, might not work for Windows...
-        #[allow(deprecated)]
-        let mut source = std::env::home_dir().expect("No home dir?");
-
-        source.push(".cargo");
-        for part in path.components().skip(ix) {
-            source.push(part);
-        }
-        if source.exists() {
-            return Some((Source::External, source));
-        } else {
-            panic!(
-                "{path:?} looks like it can be a cargo registry reference but we failed to get it"
-            );
-        }
-    }
-
-    None
 }
 
 fn load_rust_sources(
